Move video text sanitazation to utils
